+++ config/easyforms.php
@@ -2,7 +2,7 @@
 
 return [
     'display' => TRUE,
-    'form-namespace' => 'App\Http\Forms',
+    'form-namespace' => 'PlusTimeIT\EasyForms\Forms',
     'form-path' => app_path('Http/Forms'), // must not have trailing forward slash
     'routes' => [
         'examples' => TRUE,