+++ src/Controllers/Axios.php
@@ -3,14 +3,12 @@ namespace PlusTimeIT\EasyForms\Controllers;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
-use Log;
 
 class Axios extends Controller
 {
     public function load(request $request)
     {
-        $form_class = $this->form_namespace . str_replace('-', '\\', $request->form_name);
-        Log::debug(( new $form_class() )->getFields());
+        $form_class = config('easyforms.form-namespace') . '\\' . str_replace('-', '\\', $request->form_name);
         return  ( new $form_class() )->toJson();
     }
 
@@ -18,6 +16,4 @@ class Axios extends Controller
     {
         return NULL;
     }
-
-    protected $form_namespace = 'PlusTimeIT\\EasyForms\\Forms\\';
 }