+++ src/Elements/Icon.php
@@ -1,6 +1,8 @@
 <?php
 namespace PlusTimeIT\EasyForms\Elements;
 
+use PlusTimeIT\EasyForms\Traits\ConvertTrait;
+
 class Icon
 {
     public function __construct(string $icon, string $tooltip)
@@ -35,5 +37,5 @@ class Icon
 
     protected $tooltip;
 
-    use ConvertTraits;
+    use ConvertTrait;
 }