+++ src/Fields/HiddenField.php
@@ -2,14 +2,15 @@
 namespace PlusTimeIT\EasyForms\Fields;
 
 use PlusTimeIT\EasyForms\Base\EasyField;
-use PlusTimeIT\EasyForms\Traits\ConvertTraits;
+use PlusTimeIT\EasyForms\Interfaces\FieldInterface;
+use PlusTimeIT\EasyForms\Traits\{ConvertTrait, FieldTrait};
 
-class HiddenField extends EasyField
+class HiddenField extends EasyField implements FieldInterface
 {
     public function __construct(string $name, array $options = [])
     {
         $this->name = $name;
-        return $this->fillOptions($options);
+        return $this->setOptions($options);
     }
 
     protected $component = 'v-text-field';
@@ -18,7 +19,8 @@ class HiddenField extends EasyField
 
     protected $type = self::TYPE;
 
-    use ConvertTraits;
+    use ConvertTrait;
+    use FieldTrait;
 
     public const TYPE = 'hidden';
 }