+++ src/Fields/NumberField.php
@@ -2,14 +2,15 @@
 namespace PlusTimeIT\EasyForms\Fields;
 
 use PlusTimeIT\EasyForms\Base\EasyField;
-use PlusTimeIT\EasyForms\Traits\ConvertTraits;
+use PlusTimeIT\EasyForms\Interfaces\FieldInterface;
+use PlusTimeIT\EasyForms\Traits\{ConvertTrait, FieldTrait};
 
-class NumberField extends EasyField
+class NumberField extends EasyField implements FieldInterface
 {
     public function __construct(string $name, array $options = [])
     {
         $this->name = $name;
-        return $this->fillOptions($options);
+        return $this->setOptions($options);
     }
 
     public function getMax()
@@ -44,7 +45,8 @@ class NumberField extends EasyField
 
     protected $type = self::TYPE;
 
-    use ConvertTraits;
+    use ConvertTrait;
+    use FieldTrait;
 
     public const TYPE = 'number';
 }