+++ src/Traits/FieldTrait.php
@@ -98,6 +98,16 @@ trait FieldTrait
         return $this;
     }
 
+    public function setOptions(array $options): self
+    {
+        collect($options)->each(function($value, $option) {
+            if (property_exists(__CLASS__, $option)) {
+                $this->{$option} = $value;
+            }
+        });
+        return $this;
+    }
+
     public function setOrder(int $order): self
     {
         $this->order = $order;
@@ -146,5 +156,5 @@ trait FieldTrait
         return $this;
     }
 
-    use ConvertTraits;
+    use ConvertTrait;
 }